+++ lib/invalid_utf8_rejector/railtie.rb
@@ -1,7 +1,11 @@
 module InvalidUTF8Rejector
   class Railtie < Rails::Railtie
     initializer "invalid_utf8_rejector.insert_middleware" do |app|
-      app.config.middleware.insert_before 0, "InvalidUTF8Rejector::Middleware"
+      if Rails::VERSION::MAJOR >= 5
+        app.config.middleware.insert_before 0, InvalidUTF8Rejector::Middleware
+      else
+        app.config.middleware.insert_before 0, "InvalidUTF8Rejector::Middleware"
+      end
     end
   end
 end